feat: hot_or_not simulation test passes
