+++ src/canister/individual_user_template/src/api/hot_or_not_bet/receive_bet_from_bet_makers_canister.rs
@@ -1,7 +1,7 @@
 use ic_cdk_macros::update;
 use std::time::SystemTime;
 
-use candid::Principal;
+use candid::{types::internal, Principal};
 use ic_cdk::{api::management_canister::provisional::CanisterId, println};
 use shared_utils::{
     canister_specific::individual_user_template::types::{
@@ -11,6 +11,16 @@ use shared_utils::{
     },
     common::utils::system_time,
 };
+
+use shared_utils::common::types::{
+    app_primitive_type::PostId,
+    utility_token::token_event::SystemTimeInMs
+};
+use shared_utils::constant::TIMER_DURATION;
+use ic_stable_structures::btreemap::BTreeMap;
+use ic_stable_structures::{Memory, Storable};
+use std::fmt::Debug;
+use crate::api::hot_or_not_bet::tabulate_hot_or_not_outcome_for_post_slot::tabulate_hot_or_not_outcome_for_post_slot;
 
 use crate::{
     api::{
@@ -29,18 +39,24 @@ fn receive_bet_from_bet_makers_canister(
     let bet_maker_canister_id = ic_cdk::caller();
     update_last_canister_functionality_access_time();
 
-    dbg!(&place_bet_arg);
-    dbg!(&bet_maker_principal_id);
-    dbg!(&bet_maker_canister_id);
+    ic_cdk::println!("receive_bet_from_bet_makers_canister - creator_post: {:?}",&place_bet_arg.post_id);
+
+    let current_time = system_time::get_current_system_time_from_ic();
 
     let status = CANISTER_DATA.with(|canister_data_ref_cell| {
-        receive_bet_from_bet_makers_canister_impl(
-            &mut canister_data_ref_cell.borrow_mut(),
+        let canister_data = &mut canister_data_ref_cell.borrow_mut();
+       
+        let bet_result = receive_bet_from_bet_makers_canister_impl(
+            canister_data,
             &bet_maker_principal_id,
             &bet_maker_canister_id,
             place_bet_arg.clone(),
-            &system_time::get_current_system_time_from_ic(),
-        )
+            &current_time,
+        );
+
+        maybe_start_timer_based_on_bet_result(canister_data, current_time, place_bet_arg.clone(), bet_result.clone());
+
+        bet_result
     })?;
 
     CANISTER_DATA.with(|canister_data_ref_cell| {
@@ -53,6 +69,59 @@ fn receive_bet_from_bet_makers_canister(
     update_scores_and_share_with_post_cache_if_difference_beyond_threshold(&place_bet_arg.post_id);
 
     Ok(status)
+}
+
+fn maybe_start_timer_based_on_bet_result(canister_data: &mut CanisterData, current_time: SystemTime, place_bet_arg: PlaceBetArg, bet_result: Result<BettingStatus, BetOnCurrentlyViewingPostError>) {
+    // let current_time = system_time::get_current_system_time_from_ic();
+
+    if let Ok(ok_bet_result) = bet_result {
+    match ok_bet_result {
+        // this case should never match in yral game implementation
+        BettingStatus::BettingClosed => {
+            ();
+        }
+        BettingStatus::BettingOpen {
+            ongoing_slot,
+            ongoing_room,
+            ..
+        } => {
+            // insert only the first bet in first_bet_placed_at_hashmap
+            if !canister_data
+            .first_bet_placed_at_hashmap
+            .contains_key(&place_bet_arg.post_id)
+        {
+            ic_cdk::println!("current_time: {:?}", current_time);
+            canister_data.first_bet_placed_at_hashmap.insert(
+                place_bet_arg.post_id,
+                (SystemTimeInMs::from_system_time(current_time), ongoing_slot),
+            );
+            ic_cdk::println!("maybe_start_timer_based_on_bet_result: canister_data.first_bet_placed_at_hashmap: {}", print_btree_map(&canister_data.first_bet_placed_at_hashmap));
+            
+            // also push the post_id to the queue
+            let bet_timer_posts = &mut canister_data.bet_timer_posts;
+
+            let _to_print = match bet_timer_posts.insert(
+                (
+                    SystemTimeInMs::from_system_time(current_time),
+                    place_bet_arg.post_id,
+                ),
+                (),
+            ) {
+                Some(timer) => format!("Timer pushed to empty array: {:?}", timer),
+                None => "Failed to push timer to empty array".to_string(),
+            };
+
+            ic_cdk::println!(
+                "before maybe_enqueue_timer (maybe_start_timer_based_on_bet_result) - bet_timer_posts: {:?}",
+                print_btree_map(&canister_data.bet_timer_posts)
+            );
+
+            maybe_enqueue_timer(canister_data);
+        }
+        ic_cdk::println!("second bet on same post: {:?}", place_bet_arg.post_id);
+    }
+}
+    }
 }
 
 fn receive_bet_from_bet_makers_canister_impl(
@@ -71,7 +140,7 @@ fn receive_bet_from_bet_makers_canister_impl(
 
     let post = canister_data.all_created_posts.get_mut(&post_id).unwrap();
 
-    ic_cdk::println!("{:?}",&post);
+    ic_cdk::println!("receive_bet_from_bet_makers_canister_impl - post: {:?}",&post);
 
     post.place_hot_or_not_bet_v2(
         bet_maker_principal_id,
@@ -85,6 +154,157 @@ fn receive_bet_from_bet_makers_canister_impl(
         &mut canister_data.slot_details_map,
     )
 }
+
+// TIMER LOGIC STARTS HERE
+
+fn maybe_enqueue_timer(canister_data: &mut CanisterData) {
+    let should_start_timer = match canister_data.is_timer_running {
+        Some(post_id) => process_running_timer(canister_data, post_id),
+        None => !canister_data.first_bet_placed_at_hashmap.is_empty(),
+    };
+
+    ic_cdk::println!(
+        "inside maybe_enqueue_timer: timer_running = {:?} ; should_start_timer = {:?} \n",
+        &canister_data.is_timer_running,
+        should_start_timer
+    );
+
+    if should_start_timer {
+        start_timer(canister_data);
+    }
+}
+
+fn process_running_timer(canister_data: &mut CanisterData, post_id: u64) -> bool {
+    ic_cdk::println!("process_running_timer for post_id: {:?}", post_id);
+    if !timer_expired(post_id, canister_data) {
+        // don't start timer again if previous one has not expired yet
+        ic_cdk::println!("process_running_timer for post_id: {:?} -  early_return as timer is not expired", post_id);
+        return false;
+    }
+
+    ic_cdk::println!("process_running_timer - canister_data.first_bet_placed_at_hashmap for post_id: {:?}",canister_data.first_bet_placed_at_hashmap.get(&post_id));
+
+    if let Some((_, ongoing_slot)) = canister_data.first_bet_placed_at_hashmap.get(&post_id) {
+        tabulate_hot_or_not_outcome_for_post_slot(canister_data, post_id, ongoing_slot);
+
+        ic_cdk::println!(
+            "\n\n canister_data.bet_timer_posts: {:?}",
+            print_btree_map(&canister_data.bet_timer_posts)
+        );
+        ic_cdk::println!("\n\n BEFORE PROCESSING TIMER {}", "//".repeat(400));
+        // remove the post_id from the hashmap and bet_timer_posts
+        let val = canister_data.first_bet_placed_at_hashmap.remove(&post_id);
+        let same_post_id = canister_data.bet_timer_posts.pop_first();
+
+        canister_data.is_timer_running = None;
+        ic_cdk::println!(" after PROCESSING TIMER {}", "--".repeat(400));
+        ic_cdk::println!(
+            "\n\n processed_timer for -- post_id: {:?}, same_post_id: {:?}",
+            post_id,
+            same_post_id
+        );
+        ic_cdk::println!("\n\n canister_data.first_bet_placed_at_hashmap NOW: {:?}", print_btree_map(&canister_data.first_bet_placed_at_hashmap));
+        ic_cdk::println!(
+            "\n\n canister_data.bet_timer_posts: {:?}",
+            print_btree_map(&canister_data.bet_timer_posts)
+        );
+
+        // return true to indicate that timer has been processed
+        true
+    } else {
+        false
+    }
+}
+
+pub fn print_btree_map<K, V, M>(btree: &BTreeMap<K, V, M>) -> String
+where
+    K: Ord + Debug + Storable + Clone,
+    V: Debug + Storable,
+    M: Memory,
+{
+    let mut result = String::from("{");
+    let mut iter = btree.iter();
+
+    if let Some((key, value)) = iter.next() {
+        result.push_str(&format!("{:?}: {:?}", key, value));
+    }
+
+    for (key, value) in iter {
+        result.push_str(&format!(", {:?}: {:?}", key, value));
+    }
+
+    result.push('}');
+    result
+}
+
+fn start_timer(canister_data: &mut CanisterData) {
+    if !canister_data.first_bet_placed_at_hashmap.is_empty() {
+        // bet_timer_posts is a queue with head at the last element of array
+        // and tail at the first element of array.
+        // this is because `pop` removes the last entry from the vec in ic_stable_structures
+        if let Some(((insertion_time, first_post_id), _)) = canister_data.bet_timer_posts.first_key_value() {
+            // if let Some(composite_key) = canister_data.bet_timer_posts.pop_first() {
+            // ic_cdk::println!("\n--------- post_id: {}, starting timer ------\n", first_post_id);
+            ic_cdk::println!("\n--------- post_id: {:?}, starting timer ------ bet_placed_at: {:?} \n", first_post_id, insertion_time);
+            if let Some((bet_placed_time, _ongoing_slot_for_post)) =
+                canister_data.first_bet_placed_at_hashmap.get(&first_post_id)
+            {
+                let current_time = SystemTimeInMs::now();
+                let interval_for_timer = current_time.calculate_remaining_interval(&bet_placed_time, TIMER_DURATION).unwrap();
+
+                canister_data.is_timer_running = Some(first_post_id);
+
+                ic_cdk_timers::set_timer(interval_for_timer, move || {
+
+                    CANISTER_DATA.with(|canister_data_ref_cell| {
+                        let canister_data = &mut canister_data_ref_cell.borrow_mut();
+                        ic_cdk::println!("INSIDE start_timer - canister_data.first_bet_placed_at_hashmap: {}, post_id: {}, interval_for_timer: {:?} \n",print_btree_map(&canister_data.first_bet_placed_at_hashmap), first_post_id, interval_for_timer);
+
+                        maybe_enqueue_timer(canister_data);
+                    });
+                });
+            }
+        }
+    }
+}
+
+fn timer_expired(post_id: PostId, canister_data: &CanisterData) -> bool {
+    if !canister_data.first_bet_placed_at_hashmap.is_empty() {
+        if let Some(((first_time, first_post_id), _)) =
+            canister_data.bet_timer_posts.first_key_value()
+        {
+            if let Some((bet_placed_time, _ongoing_slot_for_post)) =
+                canister_data.first_bet_placed_at_hashmap.get(&post_id)
+            {
+                let current_time = SystemTimeInMs::now();
+                let interval = current_time.duration_since(&bet_placed_time);
+                let return_val = interval > TIMER_DURATION;
+
+                ic_cdk::println!(
+                    "post_id ({}) == last_post_id ({}), {}, \n bet_timer_posts: {:?}, \n bet_timer_posts.time: {:?} === first_bet_placed_at_hashmap.time: {:?}", 
+                    post_id,
+                    first_post_id,
+                    post_id == first_post_id,
+                    print_btree_map(&canister_data.bet_timer_posts),
+                    first_time, 
+                    bet_placed_time
+                );
+
+                ic_cdk::println!("post_id: {}, timer_expired: {:?}", post_id, return_val);
+
+                return return_val;
+            }
+        }
+    }
+    ic_cdk::println!(
+        "post_id: {}, timer_expired: FALSE (returning from outside)",
+        post_id
+    );
+
+    false
+}
+
+// TIMER LOGIC ENDS HERE
 
 fn update_profile_stats_with_bet_placed(
     canister_data: &mut CanisterData,
